git-svn-id: https://svn.apache.org/repos/asf/sis/trunk@1551743 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-referencing/src/main/java/org/apache/sis/geometry/AbstractDirectPosition.java
@@ -294,7 +294,7 @@ public abstract class AbstractDirectPosition implements DirectPosition {
             final char close = (c == '(') ? ')' : ']';
             final int pos = CharSequences.lastIndexOf(wkt, close, i, length);
             if (pos != --length) {
-                final int key;
+                final short key;
                 final Object[] args;
                 if (pos < 0) {
                     key  = Errors.Keys.NonEquilibratedParenthesis_2;
